+++ components/Analysis/AnalysisWrapper.tsx
@@ -2,42 +2,24 @@
 
 import { ErrorBoundary } from "react-error-boundary";
 
-import {
-  type AnalysisDataMode,
-  analysisDataModes,
-  AnalysisView,
-  analysisViews,
-} from "@/utils/types";
 import { cn } from "@/utils/cn";
-import {
-  useAnalysisActions,
-  useAnalysisDecimationThreshold,
-  useAnalysisDataMode,
-  useAnalysisSeries,
-  useAnalysisTimeLength,
-  useAnalysisView,
-} from "@/hooks/useAnalysis";
+import { useAnalysisSeries, useAnalysisTimeLength } from "@/hooks/useAnalysis";
 import { useComparisonChartQueries } from "@/hooks/useComparisonChartQueries";
 import { useUserCurrencySetting } from "@/hooks/useUserSettings";
-import { useThemeTyped } from "@/hooks/useThemeTyped";
 
 import AnalysisChart from "./AnalysisChart";
 import AnalysisLegend from "./AnalysisLegend";
+import AnalysisViewSelector from "./AnalysisViewSelector";
 import Panel from "../Theme/Panel";
 import Loader from "../Loader";
 import AnalysisRowWithProvider from "./AnalysisRow/AnalysisRow";
+import AnalysisDataSelector from "./AnalysisDataSelector";
+import AnalysisDecimationSelector from "./AnalysisDecimationSelector";
 
 const AnalysisWrapper = () => {
   const series = useAnalysisSeries();
   const timeLength = useAnalysisTimeLength();
-  const mode = useAnalysisDataMode();
-  const view = useAnalysisView();
   const currency = useUserCurrencySetting();
-  const theme = useThemeTyped();
-  const decimationThreshold = useAnalysisDecimationThreshold();
-  const thresholdOptions = [Infinity, 25, 50, 100, 150];
-  const { setDecimationThreshold, setDataMode, setView, setSeriesAxisById } =
-    useAnalysisActions();
 
   const responses = useComparisonChartQueries({
     ids: series.map((s) => s.id),
@@ -48,29 +30,6 @@ const AnalysisWrapper = () => {
   const someLoading = responses.some((r) => r.isPending);
   const allLoading = responses.every((r) => r.isPending);
   const noneSelected = series.length === 0;
-
-  function handleModeChange(e: React.ChangeEvent<HTMLInputElement>) {
-    const newMode = e.currentTarget.value as AnalysisDataMode;
-
-    if (newMode === "Rate of Return") {
-      setView("Linear");
-      series.forEach((s) => setSeriesAxisById(s.id, "left"));
-    }
-    setDataMode(newMode);
-  }
-
-  function handleViewChange(e: React.ChangeEvent<HTMLInputElement>) {
-    const newView = e.currentTarget.value as AnalysisView;
-    setView(newView);
-  }
-
-  function handleDecimationThresholdChange(
-    e: React.ChangeEvent<HTMLInputElement>
-  ) {
-    let newThreshold = Number(e.currentTarget.value);
-    newThreshold = Number.isNaN(newThreshold) ? Infinity : newThreshold;
-    setDecimationThreshold(Number(newThreshold));
-  }
 
   return (
     <div className="w-full flex flex-col items-center justify-center mb-[20vh]">
@@ -99,14 +58,7 @@ const AnalysisWrapper = () => {
           >
             <div className="h-[700px]">
               <AnalysisChart
-                series={series}
                 rawData={responses.filter((r) => r.data).map((r) => r.data!)}
-                mode={mode}
-                currency={currency}
-                theme={theme}
-                timeLength={timeLength}
-                decimationThreshold={decimationThreshold}
-                view={view}
               />
             </div>
             <div className="flex gap-x-4 mt-4">
@@ -115,61 +67,13 @@ const AnalysisWrapper = () => {
               </div>
               <div className="mr-4 w-1/2 flex justify-end gap-x-20">
                 <div>
-                  <p className="text-sm font-medium uppercase dark:text-zinc-300 text-zinc-700 mb-1">
-                    View
-                  </p>
-                  {analysisViews.map((v) => (
-                    <div key={v}>
-                      <input
-                        type="radio"
-                        id={v}
-                        value={v}
-                        className="mr-1 disabled:hover:cursor-not-allowed"
-                        disabled={mode === "Rate of Return"}
-                        checked={view === v}
-                        onChange={handleViewChange}
-                      />
-                      <label htmlFor={v}>{v}</label>
-                    </div>
-                  ))}
+                  <AnalysisViewSelector />
                 </div>
                 <div>
-                  <p className="text-sm font-medium uppercase dark:text-zinc-300 text-zinc-700 mb-1">
-                    Decimation
-                  </p>
-                  {thresholdOptions.map((opt) => (
-                    <div key={opt}>
-                      <input
-                        type="radio"
-                        id={"compression" + opt.toString()}
-                        value={opt}
-                        className="mr-1 disabled:hover:cursor-not-allowed"
-                        checked={decimationThreshold === opt}
-                        onChange={handleDecimationThresholdChange}
-                      />
-                      <label htmlFor={"compression" + opt.toString()}>
-                        {opt === Infinity ? "None" : `${opt} points`}
-                      </label>
-                    </div>
-                  ))}
+                  <AnalysisDecimationSelector />
                 </div>
                 <div>
-                  <p className="text-sm font-medium uppercase dark:text-zinc-300 text-zinc-700 mb-1">
-                    Data
-                  </p>
-                  {analysisDataModes.map((dataMode) => (
-                    <div key={dataMode}>
-                      <input
-                        type="radio"
-                        id={dataMode}
-                        value={dataMode}
-                        checked={dataMode === mode}
-                        onChange={handleModeChange}
-                        className="mr-1"
-                      />
-                      <label htmlFor={dataMode}>{dataMode}</label>
-                    </div>
-                  ))}
+                  <AnalysisDataSelector />
                 </div>
               </div>
             </div>
